refactor: split runnable registration into executor_runnables_registry.go
